refactoring in LikeService add validation that user can't remove foreign likes

+++ src/main/groovy/com/ionciolac/domain/service/LikeService.groovy
@@ -6,6 +6,7 @@ import com.ionciolac.common.util.CommonMessage
 import com.ionciolac.domain.model.Like
 import com.ionciolac.port.inputs.LikeInPort
 import com.ionciolac.port.outputs.LikeOutPort
+import org.apache.coyote.BadRequestException
 import org.springframework.stereotype.Service
 
 @Service
@@ -30,20 +31,31 @@ class LikeService implements LikeInPort {
 
     @Override
     void deleteLike(String authorizedUserId, String id) {
+        def dbLike = getDBLike(id)
+        if (dbLike.getUserId() != authorizedUserId) {
+            String msg = String.format(CommonMessage.FOREIGN_MESSAGE, CommonMessage.DELETE, CommonMessage.LIKE)
+            throw new BadRequestException(msg)
+        }
         likeOutPort.deleteLike(id)
     }
 
     @Override
     Like getLike(String authorizedUserId, String id) {
-        def dbOptionalLike = likeOutPort.getLike(id)
-        if (dbOptionalLike.isPresent())
-            return dbOptionalLike.get()
-        else
-            throw new ObjectNotFoundException(String.format(CommonMessage.NOT_FOUND_MESSAGE, CommonMessage.LIKE, id))
+        return getDBLike(id)
     }
 
     @Override
     List<Like> getPostLikes(String authorizedUserId, String id) {
         return likeOutPort.getPostLikes(id)
+    }
+
+    private def getDBLike(String id) {
+        def dbLike = likeOutPort.getLike(id)
+        if (dbLike.isPresent())
+            return dbLike.get()
+        else {
+            String msg = String.format(CommonMessage.NOT_FOUND_MESSAGE, CommonMessage.LIKE, id)
+            throw new ObjectNotFoundException(msg)
+        }
     }
 }
